Updated the format of the changelog
Because of the potential to have multiple sub-crate releases within a
single release of the main `wolf_engine` crate, I've updated the format of the
changelog to accommodate this in a more pleasing way.

Instead of doing `[crate_name x.y]`, we'll do an `h2` tag for the crate name,
and an `h3` tag for the version number.  Like this:

```
* [0.1]

- Some changes to the main `wolf_engine` crate.

** [some_sub_crate]

*** [0.2]

- Some other changes to the sub-crate.

*** [0.1]

- Some Changes to the sub-crate.

** [some_other_sub_crate]

*** [0.1]

- Some changes to some other subcrate.
```