Moved the `generic_event_queue` code into `wolf_engine_events`
Mostly to make maintaining it easier.  The code in `generic_event_queue` was
originally written as part of Wolf Engine, but I was moved out to allow it to
be shared with other crates.  The issue is it makes maintaining it pretty
annoying.

I'm also not sure the reasons for moving it out of Wolf Engine's crate are
even valid anymore, especially considering the modules are now intended to be
used independently.  I may just completely discontinue the
`generic_event_queue` package in favor of just using `wolf_engine_events`
directly.  Still not 100% on that yet, tho, so I'll probably consider it for a
while before actually doing anything about it.