Add the remaining quit-test
This really is just a smoke-test to ensure the window system actually
runs.  Most details beyond that can't really be tested.  I don't think
we really need to be testing Winit's behaviors anyways.  The examples
will need to serve as a bit of a manual test-bed to check the specifics,
and the compiler should be able to help ensure everything is hooked up
right.