Fix panic when closing windows
Pixels gets dropped before this event.  Seems the `Closed` event comes
through before the `RedrawRequested` event does.  I expected events to
come through in-order, but maybe that isn't the case.  I'm going to just
`if let` this block for now, but it'd be preferrable for the main-loop
not to send events for dropped windows in the first place.