Store windows in the context object
I think I can make this work with full `Window` objects.  If not, I can
probably store `WindowState` objects instead.

My goal here is for the `WindowContext` to do all, if not most, of it's
own bookkeeping, because that's better than having to keep multiple
handles / state objects in sync externally.